+++ CHANGELOG.md
@@ -13,7 +13,8 @@
 - The power unit of the saved raw data changes from dBm (or dB) to Watt. ([#58](https://github.com/imreburn/Wavelength-Reference/issues/58))
 - If the fields of "Pass/Fail Criteria" are left empty, 0 is assigned instead of showing an error message. ([#64](https://github.com/imreburn/Wavelength-Reference/issues/64))
 - The delta marker table is placed first below the spectrum. ([#69](https://github.com/imreburn/Wavelength-Reference/issues/69))
-- The custom peak table is shown only if the Bandwidth marker exists, above the peak table.
+- The custom peak table is shown only if the Bandwidth marker exists.
+- The custom peak table is shown right side of the delta marker table.
 - When a Bandwidth marker is placed, the (x, y) position is shown with the marker. The width is also shown next to the offset marker.
 
 ### Fixed